Introduced the Files
Brought in & organized all the images and code. Swiping images left or right int he gallery from a browser with a touch screen still needs work on the timing.

Adding a log in functionality is a priority. Eventually, this login functionality will allow the owner to add, remove, & re-order images on the gallery & homepage.