Updated & Organized Root README
Added details about what the Source Code directory contains. The header contains a link to the Source Code directory, so this is also a test to see if relative links need any slashes, or to not have spaces to work. A README will be added to that directory, but may take some time to complete given the number of files that need to be given summaries.

Also divided specific details about the website into a seperate section for clarity.