Fix/Test Head Source Code Link in README
Having a slash at the end of the link in the Source Code link in the body of the Source Code section, as well as replacing the space with its URL percent encoding (%20) fixed the relative link. This replaces the space in the head Source Code link with its URL percent encoding. The slash at the end was also removed to see if that's really necessary.