Fixed a Bug in Gallery's JS File & Compressed Several Files
There was a bug in attackhelicopter.js that caused the image to slide off the screen if the user clicked & dragged the image to the end of the screen & released it while the cursor wasn't moving.

As that file is over 52KB in size I compressed it & put it's compressed version (as well as those of many other files) in the Compressed Code directory.

The slash was also removed from the relative link in the main README as the previous commit proved that only replacing the space with its URL code (%20) was necessary, & the slash was superfluous.