Fixed a Bug in Gallery's JS File
There was a bug in attackhelicopter.js that caused the images to slide in the wrong directions if the user clicked & dragged the image to the end of the screen while keeping more than half of the held image on the screen & released it while the cursor wasn't moving. This was fixed.