A Few Fixes and Optimizations
attackhelicopter.js wasn't finding the style rule with "div#thumbnails img, div#gallery div". This isn't inherently due to the JavaScript file, but how that string in style.css and mobile.css changes when it's compressed. To fix this I changed attackhelicopter.js so that it would find the css rule with all three tokens: "#thumbnails img", ",", and "#gallery div". This is slightly less efficient, but far more flexible.

obama.js was given a line that makes the mobile version of 404NotFound.php page's navigation look better.

Every div# in style.css and mobile.css was replaced with # to make those files and their compressed versions smaller.