Added an IPLogger
Added a bit of PHP to the blacklist.php file that would log the IP address of anyone that connects, the number of times they've visited, & the port they last used to access the site.

Still need to add a session variable to prevent the number of visits from a user being incremented unless they visit the site from a new device or browser without closing the one they're already in.