Added Compatibility with Browsers that Insert Their Own Stylesheets & Made the resizeDivs() Function Far More Efficient
Made grabStyle() far more particular about which stylesheets it selects in order to make this page compatible with browsers that insert their own CSS stylesheets.

Also made resizeDivs() only run it's computations and modifications to the CSS in the case that the page changed in size enough that the number of columns that hold the images in the gallery needs to be changed.

These changes came with much improved and more detailed comments in these two functions (in the non-compressed version of attackhelicopter.js) and changing a word in the about page.