Fixed grabStyle() Bug
Made the for loop that itterates through the stylesheets itterate forwards instead of backwards so that it encounters style.css before mobile.css.

On mobile browsers, encountering mobile.css first would set some of the variables in the CSS without assigning galleryDivsStyle to the appropriate added rule in style.css.